pyproject.toml
[project]
name = "RSA_API"
version = "0.1.0"
requires-python = ">=3.11"
description = "API for RSA encryption and decryption"
readme = "README.md"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "cachecontrol==0.14.0",
    "cachetools==5.5.0",
    "certifi==2024.7.4",
    "cffi==1.17.1",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "cryptography==43.0.1",
    "debugpy==1.8.7",
    "dnspython==2.6.1",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.112.1",
    "fastapi-cli==0.0.5",
    "firebase-admin==6.5.0",
    "gcloud==0.18.3",
    "google-api-core==2.21.0",
    "google-api-python-client==2.149.0",
    "google-auth==2.35.0",
    "google-auth-httplib2==0.2.0",
    "google-cloud-core==2.4.1",
    "google-cloud-firestore==2.19.0",
    "google-cloud-storage==2.18.2",
    "google-crc32c==1.6.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.65.0",
    "greenlet==3.1.1",
    "grpcio==1.66.2",
    "grpcio-status==1.66.2",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httplib2==0.22.0",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "idna==3.7",
    "jinja2==3.1.4",
    "jwcrypto==1.5.6",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "mdurl==0.1.2",
    "msgpack==1.1.0",
    "numpy==2.1.0",
    "oauth2client==4.1.3",
    "proto-plus==1.24.0",
    "protobuf==5.28.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pycryptodome==3.21.0",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "pydantic-settings>=2.8.1",
    "pygments==2.18.0",
    "pyjwt==2.9.0",
    "pynvim==0.5.0",
    "pyparsing==3.2.0",
    "pyrebase4==4.8.0",
    "python-dotenv==1.0.1",
    "python-jwt==4.1.0",
    "python-multipart==0.0.9",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==0.10.1",
    "rich==13.7.1",
    "rsa==4.9",
    "setuptools>=78.1.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "starlette==0.38.2",
    "typer==0.12.4",
    "typing-extensions==4.12.2",
    "uritemplate==4.1.1",
    "urllib3==1.26.20",
    "uvicorn==0.30.6",
    "uvloop==0.20.0",
    "watchfiles==0.23.0",
    "websockets==13.0",
]
